[enh] some further enhancements, mainly extending the functionality of the algorithm
One particular change is that there are now three separate weights for each term, allowing different cost functions to be used in the three steps of parameter estimation. This allows us to run the classic training-estimation approach by setting the weight to 0 for the est-sum for step 1 and then setting the weight to 0 for the cal-sum for steps 2 and 3. Also, now it is possible to remove parameter constraints altogether with one single change of the transfer function (from 'sin2' to 'none').